Remove msg()
Every single line of log output is written to OutputDebugString().  This makes
setup run slowly under a  debugger.  If you have stdout connected to the
terminal where your debugger is running (e.g. 'gdb setup'), this output is
duplicated.

Examiing the other uses of msg() some of these are genuine errors, which
should be logged somewhere we might have a chance to see them.   Convert
those to LOG_PLAIN.  Convert uses of msg() for debug output to LOG_BABBLE.