Remove msg() and convert to log output
Remove msg(), which writes output to the debugger via OutputDebugString() and
convert it's uses to log output using the printf-style log adaptors.

Examining the uses of msg(), some of these are genuine errors, which should be
logged somewhere we might have a chance to see them.  Convert those to
LOG_PLAIN.  Convert the other uses of msg() for debug output to LOG_BABBLE.