Remove msg() and convert to log output
Remove msg(), which writes output to the debugger via OutputDebugString() and
convert it's uses to log output using the printf-style log adaptors.

Examining the uses of msg(), some of these are genuine errors, which should be
logged somewhere we might have a chance to see them.  Convert those to
LOG_PLAIN.  Convert the other uses of msg() for debug output to LOG_BABBLE.

ChangeLog:

2015-03-04  Jon TURNEY  <jon.turney@dronecode.org.uk>

	* msg.cc (msg): Remove.
	* msg.h (msg): Ditto.
	* compress_xz.cc (read, bid_xz, bid_lzma): Convert from msg() to
	LogBabblePrintf() or LogPlainPrintf() as appropriate.
	* crypto.cc (MESSAGE): Ditto.
	* desktop.cc (make_link, start_menu, desktop_icon)
	(check_desktop, check_startmenu): Ditto.
	* gpg-packet.cc (MESSAGE): Ditto.
	* ini.cc (do_ini_thread): Ditto.
	* simpsock.cc (SimpleSocket): Ditto.