Allow setup to parse more than 3 versions from the setup.ini file
Reminded by a recent request as to how to install xorg-server-1.17.1-2, which
has disappeared beyond setup's ken (in order to determine if there was a
regression in the curent version), this is a re-send of a patch I originally
submitted back in 2011 [1], which received an ambiguous response then.

[1] https://cygwin.com/ml/cygwin-apps/2011-04/msg00053.html

This recognizes any "[foo]" line as introducing the information for another
version, which doesn't have one of the trust levels [curr], [prev] or [test],
and so isn't automatically selected when setup is told to install all packages
at that trust level (by default, [curr]).

Setup already does all the neccessary sorting in version order etc. to use these
additional versions.

Since the value of <foo> carries no meaning, it might make sense to update the
setup.ini specification to mandate the use of specific strings like "[also]" or
"[other]", or perhaps "[prev-1]", "[prev-2]", etc.

I have written a corresponding patch to genini.

I'm not sure what expiry policy is currently used by upset for old packages, but
presumably that would need to be made more sophisticated, along with the changes
needed to generate setup.ini entries for other versions.

Signed-off-by: Jon TURNEY <jon.turney@dronecode.org.uk>