Add support for 'trigger:' lines in setup.ini
This is intended as a replacement for the setup.hint 'incver_ifdep:' line,
processed by upset, which is now only used by the _update-info-dir package.

The appearance in setup.ini of a 'trigger: <prefix>' line for a package causes
setup to create the file /var/cache/setup-triggers/<packagename> whenever a file
or directory with <prefix> as an initial substring is installed or uninstalled.

Once this is deployed, the _update-info-dir package can be updated to add
'trigger: usr/share/info/' to it's setup.hint, and contain a permanent
postinstall script which updates the info directory if the trigger file
/var/cache/setup-triggers/_update-info-dir is present.

(The current /etc/postinstall/update-info-dir script checks both /usr/info and
/usr/share/info, but there are no packages which install anything into
/usr/info, and the packaging guidelines require --infodir=/usr/share/info)

(Also currently nothing is updated when a package containing info files is
removed, which is possibly wrong)

This isn't the best solution.  Perhaps ideally, the packages which contain info
files would themselves create these trigger files on install and remove (which
could be achieved by having cygport add postinstall and preremove scripts to
create this trigger), but it's rather too late for that now.

Signed-off-by: Jon Turney <jon.turney@dronecode.org.uk>