Add menu options to expand/collapse all categories
XXX: This isn't quite right as we remember the expand/collapse request, and
reset to it when the view is refreshed.

We don't remember the expanded/collapsed state of categories, so it's not
clear what we should do when switching between views.  Perhaps reset to all
collapsed when view filte is changed?