Handle 'Source:' lines in setup.ini
This is somewhat awkward:

Storing the source package by Id rather than by name is looking a bit like a
premature optimization, but being to able to refer to source packages by Id
is very handy.

If the src pkg isn't seen until after the pkg, we don't know what it's Id
will be.  So we have to go back and fixup the Ids after we've seen all
packages.