Improve the reading of installed.db
When creating a packageversion for a package listed in installed.db,
use the information for the installed version rather than the
"current" version.  Add a new function packagedb::findBinaryVersion to
help with this.

Get as much information about an installed information as possible
from the prior reading of setup.ini.

Improve detection of installed test releases.  An installed test
release might not appear in setup.ini because it is no longer
available (e.g., it might have been replaced by a newer test release).
Detect this by comparing its version to the current version.