Add new member function SolverSolution::trans2db
This resets the package database to reflect the solver's transaction
list.  Call this if the user clicks Back on the Prerequisite page, to
allow reviewing and making changes before accepting the solver's
solution.

v2:
Drop text about seeing default solutions.  They aren't applied as we want
the user to choose there own solution.