Remove the ScanFindVisitor class
This class was used for local installs in which the repository contains no
setup.ini file.

This seems of questionable usefulness, since it doesn't know anything about
dependencies

It also doesn't correctly handle filenames with colons.

There are now good instructions at
https://cygwin.com/package-server.html#overlay for creating a valid
repository, so we just remove the class instead of trying to fix it.