Make 'System Proxy Settings' the default, rather than 'Direct'
Make 'System Proxy Settings' the default, rather than 'Direct', and re-order
the network connection options so that option is first.

If you don't need a proxy, the system proxy setting should be for direct
connection, anyhow.

So, at the moment, this is just a button you're supposed to know you need to
press to make it work, when you are behind a proxy.

This setting is persisted (as 'net-method'), so this change only effects new
installations.