Fix response to enter in the chooser search textbox
Make 'enter' after we've started typing into the search textbox cause the
search filter to immediately take effect.

We don't change the default control immediately on EN_SETFOCUS unless there
is already text in the search textbox, so pressing 'enter' with the focus on
the search textbox (the initial state of the dialog) without typing anything
into the search textbox moves to the next page, as previously.

Also improve a bit of debug output from ChooserPage::OnMessageCmd()

v2:
Explicitly add/remove the defpushbutton style from the "Next" button when
the default button changes

v3:
Refine behaviour so default control is changed on EN_SETFOCUS if some
search text is present.