Apply default solution to dependency problems
Mark the default solution in the problem report

Refactor of SolverSolution::update() so we can apply the default
solution(s).

Also:

Break out logging of the task list, so we can show it in the "dependency
problems exists, but don't use the default solution, just do what I ask"
case.

Break out 'include-source' process, so it can have effect in the case where
dependency problems exist.

v2:
Actually re-solve after adjusting the task list with default solutions.

Use the last solution as the default solution.  The solver sorts the
solutions, and I think the last one will normally be the one which removes
the problematic task from the task list