Add handling for 'replace-versions:' in setup.ini
Sometimes we make mistakes and release versions which are broken or have a
version which is wildly incorrect.  We can remove those versions from future
download, but they can linger indefinitely in installs (until a distupgrade
is done), as there is a higher version already installed.

Allow setup.ini to contain a per-package 'replace-versions:' header, which
identifies any versions which should be forcibly dist-upgraded (i.e.
downgraded to the best version in setup.ini).

(It's not possible to just instruct the solver to disfavour these specific
versions, as that's not implemented, see libsolv issue #170)