Make sure that the IniDBBuilderPackage destructor is called when needed
The IniDBBuilderPackage destructor contains code that is intended to be run
after each setup.ini file is processed.  But the IniDBBuilderPackage
variables in do_local_ini() and do_remote_ini were declared outside the loop
that processed the files.  Move the declaration inside the loop so that the
destructor is called after each iteration.