Simple keyboard accelerators for Listview
Wire-up certain actions to keyboard accelerators.

To permit selecting a subitem apparently needs us to custom-draw everything,
so for simplicity, just make the accelerators per row.

Getting the enter keypress to be delivered to the listview is tricky (the
propsheet really want to hold onto it), so for the moment we just use the
Windows menu key and space bar as accelerators.