Ensure that an installed packageversion has an ldesc if possible
In packagedb::read(), copy the ldesc from setup.ini to the
packageversion read from installed.db.  Otherwise, an installed
package with only one version will not have an ldesc to use as a
tooltip.

v2: If the installed version is no longer available, copy the ldesc
from the current version of the package.