Add double-click for a 'default action' to ListView
Add support for a double-click in the ListView to invoke a 'default
action'.

Because we receive both NM_CLICK and then NM_DBCLK, reduce the area of a
pop-up column which is sensitive to a click to the drop-down button
(which opens a focus-stealing pop-up menu), so the rest of the area can
receive a click (which does nothing) or a double-click (which does the
default action).

For a package, this default action is to toggle between the currently
installed version (or uninstalled, if not installed), and the highest
non-test version.

There is no keyboard acceleration for this action currently.