Remove 'Bin?' column
The only use this column appears to have now is that unticking it is the
same as selecting 'uninstall'.

Future work: Make 'Src?' column more independent, rather than
interacting with the 'New' (action) column in non-obvious ways. We
should be able to choose to install source irrespective of the state of
the binary package. (That should also take into account the
'--include-source' option)