Revert "Fix cases where incorrect action is stored by set_action()"
This reverts commit e435a212a6e85169616a67c08eab1d9a387c6cfc.

I'd forgotten that action = Install_action, desired = empty version is
how we represent 'install any version of this package'.  Converting that
to NoChange_action leaves us back mishandling obsolete packges specified
on the command line, fixing which was the whole point of this change. So
let's try to fix this a different way...