Ignore install action when no current package exists
This isn't ideal as a command line to install a package which only has a
test version will do nothing.  But I think we don't want installing by
category to install packages in that category which only have a test
version. Unfortunately, both cases are currently identical by the time
we get to applyCommandLinePackageSelection().