Also AttachConsole() when not elevating
Arrange to call AttachConsole() when not elevating (e.g. 'setup -B').

This seems to be needed for output to stdout to work via a Windows
pseudo-console ('conpty'), as is used with cygwin >=3.0.0.

Note that setup is a GUI application. I have no idea what it means to
write to stdout without calling AttachConsole(), but that doesn't seem
to work anymore...