Turn on __USE_MINGW_ANSI_STDIO everywhere
__USE_MINGW_ANSI_STDIO is turned on by C++ STL headers, so we're getting
this define due to that in most places, except where we happen to
include stdio.h before them. archive_tar.cc happens to do that, and
wants ANSI stdio since da31ad35, so turn it on unconditionally.

It's also turned on for C++11 by MinGW headers >= 8.0.0, which Cygwin
already has, so this already works when building there, and we're going
to have to deal with any other problem that might cause elsewhere in the
source eventually anyhow.

This fixes building since da31ad35 on F33, which only has mingw-headers
7.0.0.