Simplify uninstall-only warning
Simplify uninstall-only warning when running in 'Install from Local
Directory' mode with a non-existent package directory.

Strictly, this is wrong as written, since is_64 reflects the --arch
option setting, not the arch of the executable.

In any case, mentioning the executable name seems to just distract from
the warning we are trying to deliver here.