Simplify uninstall-only warning
Simplify uninstall-only warning when running in 'Install from Local
Directory' mode with a non-existent package directory.

As written, this is strictly wrong, since is_64 reflects the --arch
option setting, not the arch of the executable, but mentioning the
executable name seems to just distract from the warning we are trying to
deliver here.