Use package action strings from resources for column width
Use the package action strings from string table resource for computing
the column width.

(This should have been done as part of f34a20e7, but we didn't notice
the strings were also duplicated here.)

This requires a bit of enhancement to ListView::noteColumnWidth() to
work for wstrings as well.