Make '--quiet' imply 'download and install packages'
Make '--quiet' imply 'download and install packages', rather than
'whatever source mode was chosen in the last run'.  This aligns
behaviour with the help text about installation type.

This can be overriden with '--download' or '--local-install'. Specifying
both already has the meaning 'download and install packages', but let's
not require than in combination with '--quiet' to get definite
behaviour.