Don't create install package if there's no install: line
Don't erroneously create an install package if there's no install: line
for that version.

If it's for the latest version and something depends on this package,
the solver will produce a solution containing that package & version,
so we'll try to download it without having a path to fetch it from,
leading to download errors which are impossible to resolve.