Unconditionally fixup source package ids
In the case when a given package version appears in setup.ini first
with, then without an install: line, the source package is recreated by
packagemeta::add_version() to add the additional package source site,
but the install package is not - so it now contains a stale source
package id.

Unconditionally fixup source package ids, so that install package points
to the correct source package.

Also internalize libsolv attributes after doing fixup, the changes it
makes are visible (so this was probably never working right).