Show overall progress for checking packages in package cache
Show overall progress for testing presence and validating packages in
package cache.

This gives a better sense of progress when installing many packages
which are already cached (e.g. a fresh install, but all required
packagea are already in the package cache, so this step, rather than the
download one, dominates).