Add command-line option help-text localization
Change libgetopt++ Option class to store an unsigned int message catalog
identifer for the help-text, rather than a std::string of the help-text.

ParameterUsage (the function which produces the help output) now takes a
lookup function as a parameter, which translates that message identifer
to a localized std::string.

Move existing help-text to a stringtable resource, and then wire up
ParameterUsage to use LoadString.

Update libgetop++ tests appropriately.