Link with mingwex before libntll
Link with mingwex (which gcc specs will cause us to link with anyhow)
before ntll.

Some versions of the MinGW ntdll implib contain a subset of crt
functions (erroneously, they were later moved to ntllcrt).  This means
that if the linker command specifies -lntdll, an application will use
ntdll.dll for things like strnlen(), which is (a) unexpected, and (b)
causes problems, as that specific export is not provided by ntll on XP.