Drop use of std::unary_function
std::unary_function is deprecated in C++11.

visit_if() in Generic.h relies upon the callable providing an
argument_type type (which both predicate and visitor accept), so retain
that in StringConcenator, but we'd probably be better off just
open-coding visit_if(), since it's only used in the one place.