Use SetWindowLongPtrW() when setting WndProc
Using the SetWindowLongPtrA() variant means that a Unicode to ASCII shim
is inserted around the installed WndProc, which is not what we want.

(Specfically, this flattens the PropSheet caption to an 8-bit encoding
as it passes through the WndProc as a WM_SETTEXT message, which is not
reversible for non-latin scripts)

Future work: Rewrite this using SetWindowSubclass() instead, which
sidesteps all these issues.