io_stream: read/write: always return -1 on error
io_stream::read and io_stream::write are defined as ssize_t, and most
of the callers expect -1 in case of an error.

However, io_stream_file and io_stream_cygfile both use fread/fwrite and
return their return value unchanged.  Both functions return a size_t and
return 0 in both cases, EOF and error.

Fix this by checking the return value of fread and fwrite and return -1
in case of an error.

Change the two callers calling read and expecting 0 to indicate EOF and
error accordingly.