Do post-build steps in separate directories
Do post-build steps in separate directories.  This lets us keep the
uncompressed executable around to upload separately, for when problems
occur running the UPX compressed executable.

(These steps are still missing any build avoidance, but hopefully that's
not a big problem, as we generally only run them when making a release)